-Updated Formatting to match Axiom standard. - Added svn properties for line endings and token replacement

+++ Projects/Axiom/RenderSystems/OpenGLES/GLESHardwareIndexBuffer.cs
@@ -57,58 +57,56 @@ namespace Axiom.RenderSystems.OpenGLES
         int _scratchOffset;
         int _scratchSize;
 
-        public GLESHardwareIndexBuffer(HardwareBufferManager mgr,
-            IndexType idxType,int numIndexes, BufferUsage usage, bool useShadowBuffer)
-            : base(idxType, numIndexes, usage, false, useShadowBuffer)
+        public GLESHardwareIndexBuffer( HardwareBufferManager mgr, IndexType idxType, int numIndexes, BufferUsage usage, bool useShadowBuffer )
+            : base( idxType, numIndexes, usage, false, useShadowBuffer )
         {
-            if (idxType == IndexType.Size32)
+            if ( idxType == IndexType.Size32 )
             {
-                throw new AxiomException("32 bit hardware buffers are not allowed in OpenGL ES.");
+                throw new AxiomException( "32 bit hardware buffers are not allowed in OpenGL ES." );
             }
-            if (!useShadowBuffer)
+            if ( !useShadowBuffer )
             {
-                throw new AxiomException("Only support with shadowBuffer");
+                throw new AxiomException( "Only support with shadowBuffer" );
             }
-            OpenGL.GenBuffers(1, ref _bufferId);
-            GLESConfig.GlCheckError(this);
-            if (_bufferId == 0)
+            OpenGL.GenBuffers( 1, ref _bufferId );
+            GLESConfig.GlCheckError( this );
+            if ( _bufferId == 0 )
             {
-                throw new AxiomException("Cannot create GL index buffer");
+                throw new AxiomException( "Cannot create GL index buffer" );
             }
-            OpenGL.BindBuffer(All.ElementArrayBuffer, _bufferId);
-            GLESConfig.GlCheckError(this);
-            OpenGL.BufferData(All.ElementArrayBuffer, new IntPtr(sizeInBytes), IntPtr.Zero,
-                GLESHardwareBufferManager.GetGLUsage(usage));
-            GLESConfig.GlCheckError(this);
+            OpenGL.BindBuffer( All.ElementArrayBuffer, _bufferId );
+            GLESConfig.GlCheckError( this );
+            OpenGL.BufferData( All.ElementArrayBuffer, new IntPtr( sizeInBytes ), IntPtr.Zero, GLESHardwareBufferManager.GetGLUsage( usage ) );
+            GLESConfig.GlCheckError( this );
         }
+
         /// <summary>
         /// 
         /// </summary>
         protected override void UnlockImpl()
         {
-            if (_lockedToScratch)
+            if ( _lockedToScratch )
             {
-                if (_scratchUploadOnUnlock)
+                if ( _scratchUploadOnUnlock )
                 {
                     // have to write the data back to vertex buffer
-                    WriteData(_scratchOffset, _scratchSize, _scratchPtr,
-                        _scratchOffset == 0 && _scratchSize == sizeInBytes);
-                    
+                    WriteData( _scratchOffset, _scratchSize, _scratchPtr, _scratchOffset == 0 && _scratchSize == sizeInBytes );
                 }
                 // deallocate from scratch buffer
-                ((GLESHardwareBufferManager)HardwareBufferManager.Instance).DeallocateScratch(_scratchPtr);
+                ( (GLESHardwareBufferManager)HardwareBufferManager.Instance ).DeallocateScratch( _scratchPtr );
                 _lockedToScratch = false;
             }
             else
             {
-                OpenGL.BindBuffer(All.ElementArrayBuffer, _bufferId);
-                if (!OpenGLOES.UnmapBuffer(All.ElementArrayBuffer))
+                OpenGL.BindBuffer( All.ElementArrayBuffer, _bufferId );
+                if ( !OpenGLOES.UnmapBuffer( All.ElementArrayBuffer ) )
                 {
-                    throw new AxiomException("Buffer data corrupted, please reload");
+                    throw new AxiomException( "Buffer data corrupted, please reload" );
                 }
             }
             isLocked = false;
         }
+
         /// <summary>
         /// 
         /// </summary>
@@ -116,56 +114,58 @@ namespace Axiom.RenderSystems.OpenGLES
         /// <param name="length"></param>
         /// <param name="locking"></param>
         /// <returns></returns>
-        protected override IntPtr LockImpl(int offset, int length, BufferLocking locking)
+        protected override IntPtr LockImpl( int offset, int length, BufferLocking locking )
         {
-            All acces = 0;
-            if (isLocked)
+            All access = 0;
+            if ( isLocked )
             {
-                throw new AxiomException("Invalid attempt to lock an index buffer that has already been locked");
+                throw new AxiomException( "Invalid attempt to lock an index buffer that has already been locked" );
             }
 
             IntPtr retPtr = IntPtr.Zero;
-            if (length < MapBufferThreshold)
+            if ( length < MapBufferThreshold )
             {
-                retPtr = ((GLESHardwareBufferManager)HardwareBufferManager.Instance).AllocateScratch(length);
-                if (retPtr != IntPtr.Zero)
+                retPtr = ( (GLESHardwareBufferManager)HardwareBufferManager.Instance ).AllocateScratch( length );
+                if ( retPtr != IntPtr.Zero )
                 {
                     _lockedToScratch = true;
                     _scratchOffset = offset;
                     _scratchSize = length;
                     _scratchPtr = retPtr;
-                    _scratchUploadOnUnlock = (locking != BufferLocking.ReadOnly);
+                    _scratchUploadOnUnlock = ( locking != BufferLocking.ReadOnly );
 
-                    if (locking != BufferLocking.Discard)
+                    if ( locking != BufferLocking.Discard )
                     {
-                        ReadData(offset, length, retPtr);
+                        ReadData( offset, length, retPtr );
                     }//end if
                 }//end if
             }//end if
             else
             {
-                throw new AxiomException("Invalid Buffer lockSize");
+                throw new AxiomException( "Invalid Buffer lockSize" );
             }
-            if (retPtr == IntPtr.Zero)
+            if ( retPtr == IntPtr.Zero )
             {
-                OpenGL.BindBuffer(All.ElementArrayBuffer, _bufferId);
+                OpenGL.BindBuffer( All.ElementArrayBuffer, _bufferId );
                 // Use glMapBuffer
-                if (locking == BufferLocking.Discard)
+                if ( locking == BufferLocking.Discard )
                 {
-                    OpenGL.BufferData(All.ElementArrayBuffer, new IntPtr(sizeInBytes), IntPtr.Zero, GLESHardwareBufferManager.GetGLUsage(usage));
+                    OpenGL.BufferData( All.ElementArrayBuffer, new IntPtr( sizeInBytes ), IntPtr.Zero, GLESHardwareBufferManager.GetGLUsage( usage ) );
                 }
-                if ((usage & BufferUsage.WriteOnly) != 0)
-                    acces = All.WriteOnlyOes;
-
-                IntPtr pBuffer = OpenGLOES.MapBuffer(All.ElementArrayBuffer, acces);
-                if (pBuffer == IntPtr.Zero)
+                if ( ( usage & BufferUsage.WriteOnly ) != 0 )
                 {
-                    throw new AxiomException("Index Buffer: Out of memory");
+                    access = All.WriteOnlyOes;
+                }
+
+                IntPtr pBuffer = OpenGLOES.MapBuffer( All.ElementArrayBuffer, access );
+                if ( pBuffer == IntPtr.Zero )
+                {
+                    throw new AxiomException( "Index Buffer: Out of memory" );
                 }
                 unsafe
                 {
                     // return offsetted
-                    retPtr = (IntPtr)((byte*)pBuffer + offset);
+                    retPtr = (IntPtr)( (byte*)pBuffer + offset );
                 }
 
                 _lockedToScratch = false;
@@ -174,25 +174,27 @@ namespace Axiom.RenderSystems.OpenGLES
 
             return retPtr;
         }
+
         /// <summary>
         /// 
         /// </summary>
         /// <param name="offset"></param>
         /// <param name="length"></param>
         /// <param name="dest"></param>
-        public override void ReadData(int offset, int length, IntPtr dest)
+        public override void ReadData( int offset, int length, IntPtr dest )
         {
-            if (useShadowBuffer)
+            if ( useShadowBuffer )
             {
-                IntPtr srcData = shadowBuffer.Lock(offset, length, BufferLocking.ReadOnly);
-                Memory.Copy(srcData, dest, length);
+                IntPtr srcData = shadowBuffer.Lock( offset, length, BufferLocking.ReadOnly );
+                Memory.Copy( srcData, dest, length );
                 shadowBuffer.Unlock();
             }
             else
             {
-                throw new AxiomException("Reading hardware buffer is not supported");
+                throw new AxiomException( "Reading hardware buffer is not supported." );
             }
         }
+
         /// <summary>
         /// 
         /// </summary>
@@ -200,74 +202,81 @@ namespace Axiom.RenderSystems.OpenGLES
         /// <param name="length"></param>
         /// <param name="src"></param>
         /// <param name="discardWholeBuffer"></param>
-        public override void WriteData(int offset, int length, IntPtr src, bool discardWholeBuffer)
+        public override void WriteData( int offset, int length, IntPtr src, bool discardWholeBuffer )
         {
-            OpenGL.BindBuffer(All.ElementArrayBuffer, _bufferId);
-            GLESConfig.GlCheckError(this);
+            OpenGL.BindBuffer( All.ElementArrayBuffer, _bufferId );
+            GLESConfig.GlCheckError( this );
             // Update the shadow buffer
-            if (useShadowBuffer)
+            if ( useShadowBuffer )
             {
-                IntPtr destData = shadowBuffer.Lock(offset, length,
-                    discardWholeBuffer ? BufferLocking.Discard : BufferLocking.Normal);
-                Memory.Copy(src, destData, length);
+                IntPtr destData = shadowBuffer.Lock( offset, length, discardWholeBuffer ? BufferLocking.Discard : BufferLocking.Normal );
+                Memory.Copy( src, destData, length );
                 shadowBuffer.Unlock();
             }
 
-            if (offset == 0 && length == sizeInBytes)
+            if ( offset == 0 && length == sizeInBytes )
             {
-                OpenGL.BufferData(All.ElementArrayBuffer, new IntPtr(sizeInBytes), src,
-                    GLESHardwareBufferManager.GetGLUsage(usage));
-                GLESConfig.GlCheckError(this);
+                OpenGL.BufferData( All.ElementArrayBuffer, new IntPtr( sizeInBytes ), src, GLESHardwareBufferManager.GetGLUsage( usage ) );
+                GLESConfig.GlCheckError( this );
             }
             else
             {
-                if (discardWholeBuffer)
+                if ( discardWholeBuffer )
                 {
-                    OpenGL.BufferData(All.ElementArrayBuffer, new IntPtr(sizeInBytes), IntPtr.Zero,
-                    GLESHardwareBufferManager.GetGLUsage(usage));
+                    OpenGL.BufferData( All.ElementArrayBuffer, new IntPtr( sizeInBytes ), IntPtr.Zero, GLESHardwareBufferManager.GetGLUsage( usage ) );
                 }
                 // Now update the real buffer
-                OpenGL.BufferSubData(All.ElementArrayBuffer, new IntPtr(offset), new IntPtr(length), src);
-                GLESConfig.GlCheckError(this);
+                OpenGL.BufferSubData( All.ElementArrayBuffer, new IntPtr( offset ), new IntPtr( length ), src );
+                GLESConfig.GlCheckError( this );
             }
         }
+
         /// <summary>
         /// 
         /// </summary>
         protected override void UpdateFromShadow()
         {
-            if (useShadowBuffer && shadowUpdated && !suppressHardwareUpdate)
+            if ( useShadowBuffer && shadowUpdated && !suppressHardwareUpdate )
             {
-                IntPtr srcData = shadowBuffer.Lock(lockStart, lockSize, BufferLocking.ReadOnly);
-                OpenGL.BindBuffer(All.ElementArrayBuffer, _bufferId);
-                GLESConfig.GlCheckError(this);
+                IntPtr srcData = shadowBuffer.Lock( lockStart, lockSize, BufferLocking.ReadOnly );
+                OpenGL.BindBuffer( All.ElementArrayBuffer, _bufferId );
+                GLESConfig.GlCheckError( this );
 
                 // Update whole buffer if possible, otherwise normal
-                if (lockStart == 0 && lockSize == sizeInBytes)
+                if ( lockStart == 0 && lockSize == sizeInBytes )
                 {
-                    OpenGL.BufferData(All.ElementArrayBuffer, new IntPtr(sizeInBytes), srcData,
-                        GLESHardwareBufferManager.GetGLUsage(usage));
-                    GLESConfig.GlCheckError(this);
+                    OpenGL.BufferData( All.ElementArrayBuffer, new IntPtr( sizeInBytes ), srcData, GLESHardwareBufferManager.GetGLUsage( usage ) );
+                    GLESConfig.GlCheckError( this );
                 }
                 else
                 {
-                    OpenGL.BufferSubData(All.ElementArrayBuffer, new IntPtr(lockStart),
-                        new IntPtr(lockSize), srcData);
-                    GLESConfig.GlCheckError(this);
+                    OpenGL.BufferSubData( All.ElementArrayBuffer, new IntPtr( lockStart ), new IntPtr( lockSize ), srcData );
+                    GLESConfig.GlCheckError( this );
                 }
                 shadowBuffer.Unlock();
                 shadowUpdated = false;
             }
         }
+
         /// <summary>
         /// 
         /// </summary>
         /// <param name="disposeManagedResources"></param>
-        protected override void dispose(bool disposeManagedResources)
+        protected override void dispose( bool disposeManagedResources )
         {
-            OpenGL.DeleteBuffers(1, ref _bufferId);
-            GLESConfig.GlCheckError(this);
-            base.dispose(disposeManagedResources);
+            if ( !IsDisposed )
+            {
+                if ( disposeManagedResources )
+                {
+                    OpenGL.DeleteBuffers( 1, ref _bufferId );
+                    GLESConfig.GlCheckError( this );
+                }
+            }
+
+            // If it is available, make the call to the
+            // base class's Dispose(Boolean) method
+            base.dispose( disposeManagedResources );
         }
+
     }
 }
